Rotate secrets
Attempt to fix CCI

move comp logic into setup.py

Adjust cibw requirement

Update cci images

Rotate secrets

Fix secret export

Fix expired gpg keys
