+++ .github/workflows/tests.yml
@@ -17,7 +17,7 @@ jobs:
     - name: Checkout source
       uses: actions/checkout@v3
     - name: Set up Python 3.8
-      uses: actions/setup-python@v4.2.0
+      uses: actions/setup-python@v4.3.0
       with:
         python-version: 3.8
     - name: Install dependencies
@@ -40,7 +40,7 @@ jobs:
     - name: Checkout source
       uses: actions/checkout@v3
     - name: Set up Python 3.8
-      uses: actions/setup-python@v4.2.0
+      uses: actions/setup-python@v4.3.0
       with:
         python-version: 3.8
     - name: Upgrade pip
@@ -127,7 +127,7 @@ jobs:
       with:
         platforms: all
     - name: Build binary wheels
-      uses: pypa/cibuildwheel@v2.9.0
+      uses: pypa/cibuildwheel@v2.10.0
       with:
         output-dir: wheelhouse
         config-file: pyproject.toml
@@ -138,7 +138,7 @@ jobs:
       shell: bash
       run: ls -la wheelhouse
     - name: Set up Python 3.8 to combine coverage Linux
-      uses: actions/setup-python@v4.2.0
+      uses: actions/setup-python@v4.3.0
       if: runner.os == 'Linux'
       with:
         python-version: 3.8
