+++ dev/setup_secrets.sh
@@ -393,6 +393,11 @@ export_encrypted_code_signing_keys(){
 
     MAIN_GPG_KEYID=$(gpg --list-keys --keyid-format LONG "$GPG_IDENTIFIER" | head -n 2 | tail -n 1 | awk '{print $1}')
     GPG_SIGN_SUBKEY=$(gpg --list-keys --with-subkey-fingerprints "$GPG_IDENTIFIER" | grep "\[S\]" -A 1 | tail -n 1 | awk '{print $1}')
+    # Careful, if you don't have a subkey, requesting it will export more than you want.
+    # Export the main key instead (its better to have subkeys, but this is a lesser evil)
+    if [[ "$GPG_SIGN_SUBKEY" == "" ]]; then
+        GPG_SIGN_SUBKEY=$(gpg --list-keys --with-subkey-fingerprints "$GPG_IDENTIFIER" | grep "\[C\]" -A 1 | tail -n 1 | awk '{print $1}')
+    fi
     echo "MAIN_GPG_KEYID  = $MAIN_GPG_KEYID"
     echo "GPG_SIGN_SUBKEY = $GPG_SIGN_SUBKEY"
