+++ pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
-requires = ["setuptools>=41.0.1", "wheel>=0.29.0",  "Cython>=3.0.0a11"]
-# build-backend = "setuptools.build_meta"  commented out to disable pep517
+requires = ["setuptools>=41.0.1", "wheel>=0.29.0", "Cython>=3.0.0a11"]
+build-backend = "setuptools.build_meta"  # comment out to disable pep517
 
 [tool.coverage.run]
 branch = true
